Added Game_model_version.py and Model.py
Model.py implements a simple first generation model for the game.
Game_model_version.py is very similar to Game.py but is modified to take a model as an argument which plays the game instead of reading keyboard inputs. 

Model.py must be run from the same folder as Game_model_version.py otherwise i'm pretty sure the import will not work.